[REF] web,*: move barcode logic from webclient to core
This commit fixes a dependency issue by moving the barcode logic
(dialog, scanner...) from webclient/ to core/. Indeed, Many2OneField
uses it, but ideally, the views/ folder should not have dependencies
to the webclient/ folder. That dependency showed that the barcode
logic wasn't defined at the right place.

Part of task 3446824
