+++ addons/barcodes/static/src/components/barcode_scanner.js
@@ -1,8 +1,8 @@
 /** @odoo-module **/
 
 import { _t } from "@web/core/l10n/translation";
-import { scanBarcode } from "@web/webclient/barcode/barcode_dialog";
-import { isBarcodeScannerSupported } from "@web/webclient/barcode/barcode_video_scanner";
+import { scanBarcode } from "@web/core/barcode/barcode_dialog";
+import { isBarcodeScannerSupported } from "@web/core/barcode/barcode_video_scanner";
 import { Component } from "@odoo/owl";
 import { useService } from "@web/core/utils/hooks";
