+++ addons/hr_attendance/__manifest__.py
@@ -75,16 +75,6 @@ actions(Check in/Check out) performed by them.
             "web/static/src/views/fields/formatters.js",
 
             # Barcode reader utils
-            "web/static/src/webclient/barcode/barcode_dialog.js",
-            "web/static/src/webclient/barcode/barcode_dialog.xml",
-            "web/static/src/webclient/barcode/barcode_dialog.scss",
-            "web/static/src/webclient/barcode/barcode_video_scanner.js",
-            "web/static/src/webclient/barcode/barcode_video_scanner.xml",
-            "web/static/src/webclient/barcode/crop_overlay.js",
-            "web/static/src/webclient/webclient_layout.scss",
-            "web/static/src/webclient/barcode/crop_overlay.xml",
-            "web/static/src/webclient/barcode/crop_overlay.scss",
-            "web/static/src/webclient/barcode/ZXingBarcodeDetector.js",
             "barcodes/static/src/components/barcode_scanner.js",
             "barcodes/static/src/components/barcode_scanner.xml",
             "barcodes/static/src/components/barcode_scanner.scss",
