+++ addons/web/static/src/core/barcode/ZXingBarcodeDetector.js
