+++ addons/web/static/src/core/barcode/barcode_dialog.scss
