+++ addons/web/static/src/core/barcode/barcode_video_scanner.xml
