+++ addons/web/static/src/core/barcode/crop_overlay.scss
